backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLModel==0.0.16
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
alembic==1.12.1